Further Restructuring
It seems that the files in Common need to be included in tools from the get-go, otherwise the auto-updater (chocopkgup) doesn't support the package (since it copies package contents to another folder and stuff outside the package folder is not included.

Since I don't have more than one package now, its no big deal, but if it changes again, I will have to figure out a solution to sharing.